+++ lib/compat/wordpress-6.0/class-wp-rest-block-pattern-categories-controller.php
@@ -136,13 +136,13 @@ class WP_REST_Block_Pattern_Categories_Controller extends WP_REST_Controller {
 					'description' => __( 'The category name.', 'gutenberg' ),
 					'type'        => 'string',
 					'readonly'    => true,
-					'context'     => array( 'view', 'embed' ),
+					'context'     => array( 'view', 'edit', 'embed' ),
 				),
 				'label' => array(
 					'description' => __( 'The category label, in human readable format.', 'gutenberg' ),
 					'type'        => 'string',
 					'readonly'    => true,
-					'context'     => array( 'view', 'embed' ),
+					'context'     => array( 'view', 'edit', 'embed' ),
 				),
 			),
 		);