+++ phpunit/block-supports/typography-test.php
@@ -7,10 +7,26 @@
  */
 
 class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
+	/**
+	 * @var string|null
+	 */
+	private $test_block_name;
+
+	function setUp() {
+		parent::setUp();
+		$this->test_block_name = null;
+	}
+
+	function tearDown() {
+		unregister_block_type( $this->test_block_name );
+		$this->test_block_name = null;
+		parent::tearDown();
+	}
 
 	function test_font_size_slug_with_numbers_is_kebab_cased_properly() {
+		$this->test_block_name = 'test/font-size-slug-with-numbers';
 		register_block_type(
-			'test/font-size-slug-with-numbers',
+			$this->test_block_name,
 			array(
 				'api_version' => 2,
 				'attributes'  => array(
@@ -26,7 +42,7 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 			)
 		);
 		$registry   = WP_Block_Type_Registry::get_instance();
-		$block_type = $registry->get_registered( 'test/font-size-slug-with-numbers' );
+		$block_type = $registry->get_registered( $this->test_block_name );
 
 		$block_atts = array( 'fontSize' => 'h1' );
 
@@ -34,13 +50,12 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 		$expected = array( 'class' => 'has-h-1-font-size' );
 
 		$this->assertSame( $expected, $actual );
-		unregister_block_type( 'test/font-size-slug-with-numbers' );
 	}
 
 	function test_font_family_with_legacy_inline_styles_using_a_value() {
-		$block_name = 'test/font-family-with-inline-styles-using-value';
+		$this->test_block_name = 'test/font-family-with-inline-styles-using-value';
 		register_block_type(
-			$block_name,
+			$this->test_block_name,
 			array(
 				'api_version' => 2,
 				'attributes'  => array(
@@ -56,20 +71,19 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 			)
 		);
 		$registry   = WP_Block_Type_Registry::get_instance();
-		$block_type = $registry->get_registered( $block_name );
+		$block_type = $registry->get_registered( $this->test_block_name );
 		$block_atts = array( 'style' => array( 'typography' => array( 'fontFamily' => 'serif' ) ) );
 
 		$actual   = gutenberg_apply_typography_support( $block_type, $block_atts );
 		$expected = array( 'style' => 'font-family: serif;' );
 
 		$this->assertSame( $expected, $actual );
-		unregister_block_type( $block_name );
 	}
 
 	function test_typography_with_skipped_serialization_block_supports() {
-		$block_name = 'test/typography-with-skipped-serialization-block-supports';
+		$this->test_block_name = 'test/typography-with-skipped-serialization-block-supports';
 		register_block_type(
-			$block_name,
+			$this->test_block_name,
 			array(
 				'api_version' => 2,
 				'attributes'  => array(
@@ -89,7 +103,7 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 			)
 		);
 		$registry   = WP_Block_Type_Registry::get_instance();
-		$block_type = $registry->get_registered( $block_name );
+		$block_type = $registry->get_registered( $this->test_block_name );
 		$block_atts = array(
 			'style' => array(
 				'typography' => array(
@@ -105,13 +119,12 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 		$expected = array();
 
 		$this->assertSame( $expected, $actual );
-		unregister_block_type( $block_name );
 	}
 
 	function test_letter_spacing_with_individual_skipped_serialization_block_supports() {
-		$block_name = 'test/letter-spacing-with-individua-skipped-serialization-block-supports';
+		$this->test_block_name = 'test/letter-spacing-with-individua-skipped-serialization-block-supports';
 		register_block_type(
-			$block_name,
+			$this->test_block_name,
 			array(
 				'api_version' => 2,
 				'attributes'  => array(
@@ -130,20 +143,19 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 			)
 		);
 		$registry   = WP_Block_Type_Registry::get_instance();
-		$block_type = $registry->get_registered( $block_name );
+		$block_type = $registry->get_registered( $this->test_block_name );
 		$block_atts = array( 'style' => array( 'typography' => array( 'letterSpacing' => '22px' ) ) );
 
 		$actual   = gutenberg_apply_typography_support( $block_type, $block_atts );
 		$expected = array();
 
 		$this->assertSame( $expected, $actual );
-		unregister_block_type( $block_name );
 	}
 
 	function test_font_family_with_legacy_inline_styles_using_a_css_var() {
-		$block_name = 'test/font-family-with-inline-styles-using-css-var';
+		$this->test_block_name = 'test/font-family-with-inline-styles-using-css-var';
 		register_block_type(
-			$block_name,
+			$this->test_block_name,
 			array(
 				'api_version' => 2,
 				'attributes'  => array(
@@ -159,20 +171,19 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 			)
 		);
 		$registry   = WP_Block_Type_Registry::get_instance();
-		$block_type = $registry->get_registered( $block_name );
+		$block_type = $registry->get_registered( $this->test_block_name );
 		$block_atts = array( 'style' => array( 'typography' => array( 'fontFamily' => 'var:preset|font-family|h1' ) ) );
 
 		$actual   = gutenberg_apply_typography_support( $block_type, $block_atts );
 		$expected = array( 'style' => 'font-family: var(--wp--preset--font-family--h-1);' );
 
 		$this->assertSame( $expected, $actual );
-		unregister_block_type( $block_name );
 	}
 
 	function test_font_family_with_class() {
-		$block_name = 'test/font-family-with-class';
+		$this->test_block_name = 'test/font-family-with-class';
 		register_block_type(
-			$block_name,
+			$this->test_block_name,
 			array(
 				'api_version' => 2,
 				'attributes'  => array(
@@ -188,14 +199,12 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 			)
 		);
 		$registry   = WP_Block_Type_Registry::get_instance();
-		$block_type = $registry->get_registered( $block_name );
+		$block_type = $registry->get_registered( $this->test_block_name );
 		$block_atts = array( 'fontFamily' => 'h1' );
 
 		$actual   = gutenberg_apply_typography_support( $block_type, $block_atts );
 		$expected = array( 'class' => 'has-h-1-font-family' );
 
 		$this->assertSame( $expected, $actual );
-		unregister_block_type( $block_name );
 	}
-
 }